Refactor to use addTracks Function
Not totally sure if this is the right way just yet so I’ve branched it
off. Long term goal would be able to safely load each object
individually and delete them so refactoring these to be separate events
and consume the same methods is important.

addTrack is now looking to see if the this.tick object exists and if it
doesn’t it will create that object and the new tick points otherwise it
will just add the tick points with .addTickPoints.

Also created a new function to add the clock and removed from initialize